fix(settings): 設定UIの改善 (close #28)
- パネルタイトル(h2)を削除してサイドバーボタン名との重複を解消
- プライバシー説明文を「アカウントの公開設定を変更します」に変更
- フォローリクエスト: 0件の場合はメニューから非表示
- 「ミュートワード」「ルームミュート」を「ミュート設定」として統合
  /settings/mutes にワード/アニメのタブを追加
- /settings/rooms/mutes を /settings/mutes?tab=anime にリダイレクト
- ルームセクションから「ミュート」項目を削除
